move channels structs to globalplatform pkg
